requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pymongo>=4.6.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
flask-jwt-extended>=4.6.0
